refactor(v4): merge the two maths modules into src/utils/maths.ts
There were two, and the split had stopped meaning anything. `src/math.ts`
held the decay and inertia the drag service needs; `migration/utils/math.ts`
held v3's range helpers plus a `damp()` that imported from the other one. So
the dependency already crossed the boundary, and the boundary was "which
file did it arrive in", not what the code is for.

One module now, ordered by what the reader wants: the reference frame and the
default, the range helpers, then decay, then inertia. The two spec files
merge the same way.

Nothing changes behaviourally — same functions, same signatures. `migration`
keeps re-exporting them so the ported components' surface is unchanged, and
git tracked both moves as renames so the history follows.
