+++ packages/v4/migration/index.ts
@@ -17,6 +17,6 @@ export * from './Transition/index.js';
 export * from './utils/easings.js';
 export * from './utils/focus.js';
 export * from './utils/keyframes.js';
-export * from './utils/math.js';
+export * from '../src/utils/maths.js';
 export * from './utils/transition.js';
 export * from './utils/uid.js';
